change nextInt to Integer.toUnsignedLong().
Technically not needed for this project's implementation. However, this will give expected results if this code was ever used elsewhere. IE nextInt() and nextInt(n) return effectively a uint32. This has to be done by wrapping the value in a long. Such is the price of Java's signed-only primitives.

+++ src/splendor/common/cards/Deck.java
@@ -23,10 +23,10 @@ public class Deck {
 		int numCards = cards.size();
 
 		for(int i = 0; i < numCards - 2; i++){
-			//j is a random integer such that i ≤ j < n
-			int j = rand.nextInt(numCards);
+			//j is a random integer such that i ≤ j < n, for our purposes this is always a positive int
+			int j = (int) rand.nextInt(numCards);
 			while(j < i)
-				j = rand.nextInt(numCards);
+				j = (int) rand.nextInt(numCards);
 			temp = cards.get(i);
 			cards.set(i, cards.get(j));
 			cards.set(j, temp);
